pyproject.toml
[project]
name = "maira-interp"
version = "0.1.0"
description = "MAIRA-2 VLM Attention Visualization Tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.48.0,<4.52",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.64.0",
    "requests>=2.28.0",
    "sentencepiece>=0.1.97",
    "protobuf>=3.20.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
    "accelerate>=0.20.0",
    "packaging>=21.0",
]
